refactor(ingestion): move Bluesky client logic into BlueskyClient
- Create BlueskyClient in data_platform/ingestion/integrations/bluesky.py
- Move init_bluesky_client, search/fetch helpers, and row mapping into the client
- Return BlueskyFetchResult from fetch_posts_for_keyword
- Remove Bluesky-specific code from sync_clients.py
- Refactor sync_bluesky.py to orchestrate via BlueskyClient
- Update tests and architecture runbook to match the new client
